-Added Hilt Dependency Injection -Created MyApplication.kt app class -Created LoginFragment.kt for @AndroidEntryPoint -Started refactoring with Clean Architecture
