+++ .idea/misc.xml
@@ -20,6 +20,8 @@
         <entry key="..\:/Users/komza/AndroidStudioProjects/IslamicGlobalMessenger/app/src/main/res/layout/fragment_user_profile.xml" value="0.17727272727272728" />
         <entry key="..\:/Users/komza/AndroidStudioProjects/IslamicGlobalMessenger/app/src/main/res/layout/item_chats.xml" value="0.264" />
         <entry key="..\:/Users/komza/AndroidStudioProjects/IslamicGlobalMessenger/app/src/main/res/layout/item_contactview.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.10885416666666667" />
+        <entry key="app/src/main/res/layout/fragment_login.xml" value="0.10885416666666667" />
       </map>
     </option>
   </component>
