+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/data/models/UserData.kt
@@ -1,4 +1,4 @@
-package com.limuealimi.islamicglobalmessengerfromyasimin.models
+package com.limuealimi.islamicglobalmessengerfromyasimin.data.models
 
 data class UserData(
     val count: Int,
