+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/data/network/ApiClient.kt
@@ -1,11 +1,10 @@
-package com.limuealimi.islamicglobalmessengerfromyasimin.retrofit
+package com.limuealimi.islamicglobalmessengerfromyasimin.data.network
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.create
-import java.util.*
 
 object ApiClient {
+
     const val BASE_URL = "https://api.genderize.io"
 
     //?name=peter&apikey=edfcd65ad93fcde6e1efac9bc71c3a13
