+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/di/MyApplication.kt
@@ -1,0 +1,8 @@
+package com.limuealimi.islamicglobalmessengerfromyasimin.di
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication : Application() {
+}
