--- app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/DiscoverFragment.kt
@@ -1,16 +1,0 @@
-package com.limuealimi.islamicglobalmessengerfromyasimin.ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.limuealimi.islamicglobalmessengerfromyasimin.R
-import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentDiscoverBinding
-import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentUserProfileBinding
-import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
-
-class DiscoverFragment : Fragment(R.layout.fragment_discover) {
-
-    private val binding by viewBinding(FragmentDiscoverBinding::bind)
-}
