--- app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/RecordsFragment.kt
@@ -1,15 +1,0 @@
-package com.limuealimi.islamicglobalmessengerfromyasimin.ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.limuealimi.islamicglobalmessengerfromyasimin.R
-import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentRecordsBinding
-import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentUserProfileBinding
-import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
-
-class RecordsFragment : Fragment(R.layout.fragment_records) {
-    private val binding by viewBinding(FragmentRecordsBinding::bind)
-}
