--- app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/StoriesFragment.kt
@@ -1,15 +1,0 @@
-package com.limuealimi.islamicglobalmessengerfromyasimin.ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.limuealimi.islamicglobalmessengerfromyasimin.R
-import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentStoriesBinding
-import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentUserProfileBinding
-import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
-
-class StoriesFragment : Fragment(R.layout.fragment_stories) {
-    private val binding by viewBinding(FragmentStoriesBinding::bind)
-}
