+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/authenticating/LoginFragment.kt
@@ -1,0 +1,29 @@
+package com.limuealimi.islamicglobalmessengerfromyasimin.ui.authenticating
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.FragmentLoginBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LoginFragment : Fragment() {
+    private var _binding: FragmentLoginBinding? = null
+    private val binding
+        get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentLoginBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+}
