+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/discover/DiscoverFragment.kt
@@ -1,0 +1,18 @@
+package com.limuealimi.islamicglobalmessengerfromyasimin.ui.discover
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.limuealimi.islamicglobalmessengerfromyasimin.R
+
+class DiscoverFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_discover, container, false)
+    }
+}
