+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/main/MainActivity.kt
@@ -1,0 +1,23 @@
+package com.limuealimi.islamicglobalmessengerfromyasimin.ui.main
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.Navigation
+import com.limuealimi.islamicglobalmessengerfromyasimin.R
+import com.limuealimi.islamicglobalmessengerfromyasimin.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+    lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+
+    override fun onNavigateUp(): Boolean {
+        return Navigation.findNavController(this, R.id.my_nav_host_fragment).navigateUp()
+    }
+}
