+++ app/src/main/java/com/limuealimi/islamicglobalmessengerfromyasimin/ui/records/RecordsFragment.kt
@@ -1,0 +1,18 @@
+package com.limuealimi.islamicglobalmessengerfromyasimin.ui.records
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.limuealimi.islamicglobalmessengerfromyasimin.R
+
+class RecordsFragment : Fragment(R.layout.fragment_records) {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_records, container, false)
+    }
+}
