+++ app/src/main/res/layout/fragment_chats.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.ChatsFragment">
+    tools:context=".ui.chats.ChatsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/chatsRv"
