+++ app/src/main/res/layout/fragment_discover.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.DiscoverFragment">
+    tools:context=".ui.discover.DiscoverFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
