+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:orientation="vertical"
     android:padding="30dp"
-    tools:context=".ui.HomeFragment">
+    tools:context=".ui.home.HomeFragment">
 
 
     <com.google.android.material.textfield.TextInputLayout
