+++ app/src/main/res/layout/fragment_records.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.RecordsFragment">
+    tools:context=".ui.records.RecordsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
