+++ app/src/main/res/layout/fragment_stories.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.StoriesFragment">
+    tools:context=".ui.stories.StoriesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
